refactor(mm-preprocessor): rename pipeline -> processor, par -> execution
Post-boundary cleanup: with the driver gone there is no pipeline in the
crate — the module holds the family interface, so it is `processor`
(ProcessorSpec / build_processor / processor_from_spec follow). `par` was
both terse and inaccurate (without the `parallel` feature it is not
parallel at all); `execution` names what it decides — how the kernels
execute (pool vs inline) — matching llm-multimodal's execution.rs
precedent.
